create MovieDatabase
