+++ challangeNetflix/MovieAux.swift
@@ -26,4 +26,6 @@ class MovieAux {
         
         return moviesAux
     }
+    
+    
 }
